refactor: isolate inventory archive readers
